fixed remaining tests
