+++ tests/util_tests/test_shotgun.py
@@ -448,84 +448,3 @@ class TestCalcPathCache(TankTestBase):
         root_name, path_cache = tank.util.shotgun._calc_path_cache(self.tk, input_path)
         self.assertEqual("primary", root_name)
         self.assertEqual(expected, path_cache)
-
-
-class TestCreateSessionBasedConnection(TankTestBase):
-    """
-    Tests the creation of a session based Shotgun connection.
-    """
-
-    def setUp(self):
-        pass
-
-    def tearDown(self):
-        pass
-
-    @patch("tank_vendor.shotgun_authentication.session_cache.AuthenticationManager.get_connection_information")
-    @patch("tank_vendor.shotgun_authentication.connection.create_or_renew_sg_connection_from_session")
-    def test_no_script_user_uses_human_user(
-        self,
-        create_or_renew_sg_connection_from_session_mock,
-        get_connection_information_mock
-    ):
-        """
-        Makes sure having no user configured in shotgun.yml will invoke the
-        create_or_renew_sg_connection_from_session function.
-        """
-        config_data = {
-            "host": "https://something.shotgunstudio.com",
-            "login": "some_login",
-            "session_token": "session_token"
-        }
-        get_connection_information_mock.return_value = config_data
-        create_or_renew_sg_connection_from_session_mock.return_value = mockgun.Shotgun(
-            config_data["host"]
-        )
-
-        # Make sure the engine is None otherwise failing to authenticate (something.shotgunstudio.com doesn't exist)
-        # would trigger ui prompts.
-        self.assertIsNone(tank.platform.engine.current_engine())
-        # Because the credentials are invalid and there is no engine, no connection can be created.
-        tank.util.shotgun.create_sg_connection()
-        # Make sure there was an attempt to authenticate using cache session info.
-        self.assertEqual(create_or_renew_sg_connection_from_session_mock.call_count, 1)
-
-    @patch("tank_vendor.shotgun_authentication.connection.create_sg_connection_from_session")
-    @patch("tank_vendor.shotgun_authentication.interactive_authentication.renew_session")
-    def test_create_connection_with_session_renewal(self, renew_session_mock, create_sg_connection_from_session_mock):
-        """
-        When there is no valid session cached, the engine's renew session should take care of the
-        session renewal
-        """
-
-        new_connection = mockgun.Shotgun("https://something.shotgunstudio.com")
-        # First call will fail creating something from the cache, and the second call will be the 
-        # after we renwed the session.
-        create_sg_connection_from_session_mock.side_effect = [None, new_connection]
-        renew_session_mock.return_value = None
-
-        result = connection.create_or_renew_sg_connection_from_session(
-            "unused"
-        )
-
-        # Make sure we tried to renew the sesion
-        self.assertTrue(renew_session_mock.called)
-        self.assertEqual(create_sg_connection_from_session_mock.call_count, 2)
-        self.assertEqual(id(result), id(new_connection))
-
-    @patch("tank_vendor.shotgun_authentication.connection.create_sg_connection_from_session")
-    @patch("tank_vendor.shotgun_authentication.interactive_authentication.renew_session")
-    def test_create_connection_with_session_renewal_failure(self, renew_session_mock, create_sg_connection_from_session_mock):
-        """
-        When there is no valid session cached, the engine's renew session should take care of the
-        session renewal, but if the session renewal failed, we should get a TankAuthenticationError
-        """
-        # Always fail creating a cached session
-        create_sg_connection_from_session_mock.return_value = None
-        renew_session_mock.return_value = None
-
-        with self.assertRaisesRegexp(auth_errors.AuthenticationError, "failed"):
-            connection.create_or_renew_sg_connection_from_session("unused")
-        # Make sure we tried to renew the sesion
-        self.assertTrue(renew_session_mock.called)
-        self.assertEqual(create_sg_connection_from_session_mock.call_count, 2)
